feat: bundle infracost CLI into the extension
This change bundles Infracost CLI into the extension path. This means users do not have to separately install the binary.
Furthermore it means that we can scale versions of vscode with the CLI without worrying if users have an incompatible CLI version.
Additionally this allows for an easier registration process, keeping everything within the VSCode UI.

This PR also refactors the structure of the extension to be easier to work with.
