+++ .github/workflows/release.yml
@@ -11,6 +11,27 @@ permissions:
 jobs:
   build:
     name: Release
+    strategy:
+      matrix:
+        include:
+          - vsce_target: win32-x64
+            bin_target: windows-amd64
+            npm_config_arch: x64
+          - vsce_target: win32-arm64
+            ls_target: windows-arm64
+            npm_config_arch: arm
+          - vsce_target: linux-x64
+            ls_target: linux-amd64
+            npm_config_arch: x64
+          - vsce_target: linux-arm64
+            ls_target: linux-arm64
+            npm_config_arch: arm64
+          - vsce_target: darwin-x64
+            bing_target: darwin-amd64
+            npm_config_arch: x64
+          - vsce_target: darwin-arm64
+            bin_target: darwin-arm64
+            npm_config_arch: arm64
     runs-on: ubuntu-latest
     steps:
       - name: Checkout code
@@ -18,17 +39,34 @@ jobs:
       - name: Setup node
         uses: actions/setup-node@v3
         with:
-          node-version: "14"
+          node-version: "16.13.1"
       - name: Setup Environment
         run: node -e "console.log('PACKAGE_VERSION=' + require('./package.json').version + '\nPACKAGE_NAME=' + require('./package.json').name + '-' + require('./package.json').version)" >> $GITHUB_ENV
       - name: Verify versions
         run: node -e "if ('refs/tags/v' + '${{ env.PACKAGE_VERSION }}' !== '${{ github.ref }}') { console.log('::error' + 'Version Mismatch. refs/tags/v' + '${{ env.PACKAGE_VERSION }}', '${{ github.ref }}'); throw Error('Version Mismatch')} "
       - name: Install
         run: yarn
+        env:
+          npm_config_arch: ${{ matrix.npm_config_arch }}
       - name: Package extension
         run: yarn vscode:package
+        env:
+          INFRACOST_BIN_TARGET: ${{ matrix.bin_target }}
+      - name: Upload vsix as artifact
+        uses: actions/upload-artifact@v3
+        with:
+          name: ${{ matrix.vsce_target }}
+          path: "*.vsix"
+
+  publish:
+    name: Publish All
+    runs-on: ubuntu-latest
+    needs: build
+    if: success() && startsWith( github.ref, 'refs/tags/v')
+    steps:
+      - uses: actions/download-artifact@v3
       - name: Publish Extension
-        run: yarn vsce publish --yarn --packagePath ./${{ env.PACKAGE_NAME }}.vsix -p ${{ secrets.MARKETPLACE_PAT }}
+        run: yarn vsce publish --yarn --packagePath $(find . -iname *.vsix) -p ${{ secrets.MARKETPLACE_PAT }}
       - name: Create GitHub release
         id: create_release
         uses: softprops/action-gh-release@v1
