+++ .vscodeignore
@@ -2,6 +2,7 @@
 node_modules
 out/
 src/
+scripts/
 tsconfig.json
 webpack.config.js
 .github
