+++ package.json
@@ -1,117 +1,135 @@
 {
-	"name": "infracost",
-	"displayName": "Infracost",
-	"description": "Cloud cost estimates for Terraform in your editor",
-	"version": "0.1.16",
-	"publisher": "Infracost",
-	"license": "Apache-2.0",
-	"icon": "infracost-logo.png",
-	"engines": {
-		"vscode": "^1.67.0"
-	},
-	"categories": [
-		"Other",
-		"Formatters",
-		"Linters"
-	],
-	"keywords": [
-		"devops",
-		"terraform",
-		"hcl"
-	],
-	"extensionDependencies": [
-		"HashiCorp.terraform"
-	],
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/infracost/vscode-infracost.git"
-	},
-	"activationEvents": [
-		"workspaceContains:**/*.tf"
-	],
-	"main": "./dist/extension.js",
-	"contributes": {
-		"viewsContainers": {
-			"activitybar": [
-				{
-					"id": "infracost-projects",
-					"title": "Infracost",
-					"icon": "media/infracost.svg"
-				}
-			]
-		},
-		"views": {
-			"infracost-projects": [
-				{
-					"id": "infracostProjects",
-					"name": "Projects overview",
-					"icon": "media/infracost.svg",
-					"contextualTitle": "Infracost Projects"
-				}
-			]
-		},
-		"commands": [
-			{
-				"command": "infracost.resourceBreakdown",
-				"title": "Show all the cost components for a given resource"
-			},
-			{
-				"command": "infracost.refresh",
-				"title": "Refresh",
-				"icon": {
-					"light": "resources/light/refresh.svg",
-					"dark": "resources/dark/refresh.svg"
-				}
-			}
-		],
-		"menus": {
-			"view/title": [
-				{
-					"command": "infracost.refresh",
-					"when": "view == infracostProjects",
-					"group": "navigation"
-				}
-			],
-			"view/item/context": [ ]
-		}
-	},
-	"scripts": {
-		"vscode:package": "vsce package --yarn",
-		"vscode:prepublish": "yarn package",
-		"compile": "webpack",
-		"watch": "webpack --watch",
-		"package": "webpack --mode production --devtool hidden-source-map",
-		"lint": "eslint src --ext ts,js --ignore-path .eslintignore --ignore-path .gitignore . --max-warnings=0",
-		"lint:fix": "eslint src --fix --ext ts,js --ignore-path .eslintignore --ignore-path .gitignore . ",
-		"format": "prettier --write 'src/**/*.{js,ts}'",
-		"format:check": "prettier --check 'src/**/*.{js,ts}'"
-	},
-	"devDependencies": {
-		"@types/glob": "^7.2.0",
-		"@types/mocha": "^9.1.1",
-		"@types/node": "14.x",
-		"@types/vscode": "^1.67.0",
-		"@typescript-eslint/eslint-plugin": "^5.30.5",
-		"@typescript-eslint/parser": "^5.30.5",
-		"@vscode/test-electron": "^2.1.3",
-		"eslint": "^8.14.0",
-		"eslint-config-airbnb-base": "^15.0.0",
-		"eslint-config-prettier": "^8.5.0",
-		"eslint-plugin-import": "^2.26.0",
-		"eslint-plugin-prettier": "^4.2.1",
-		"file-loader": "^6.2.0",
-		"glob": "^8.0.1",
-		"mocha": "^9.2.2",
-		"prettier": "^2.7.1",
-		"ts-loader": "^9.2.8",
-		"typescript": "^4.6.4",
-		"vsce": "^2.9.2",
-		"webpack": "^5.70.0",
-		"webpack-cli": "^4.9.2"
-	},
-	"dependencies": {
-		"@types/semver": "^7.3.12",
-		"handlebars": "^4.7.7",
-		"semver": "^7.3.8"
-	}
+  "name": "infracost",
+  "displayName": "Infracost",
+  "description": "Cloud cost estimates for Terraform in your editor",
+  "version": "0.1.16",
+  "publisher": "Infracost",
+  "license": "Apache-2.0",
+  "icon": "infracost-logo.png",
+  "engines": {
+    "vscode": "^1.67.0"
+  },
+  "categories": [
+    "Other",
+    "Formatters",
+    "Linters"
+  ],
+  "keywords": [
+    "devops",
+    "terraform",
+    "hcl"
+  ],
+  "extensionDependencies": [
+    "HashiCorp.terraform"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/infracost/vscode-infracost.git"
+  },
+  "activationEvents": [
+    "workspaceContains:**/*.tf"
+  ],
+  "main": "./dist/extension.js",
+  "contributes": {
+    "viewsContainers": {
+      "activitybar": [
+        {
+          "id": "infracost-projects",
+          "title": "Infracost",
+          "icon": "media/infracost.svg"
+        }
+      ]
+    },
+    "views": {
+      "infracost-projects": [
+        {
+          "id": "infracostAuth",
+          "name": "Authenticate",
+          "when": "!infracost:loggedIn"
+        },
+        {
+          "id": "infracostProjects",
+          "name": "Projects overview",
+          "icon": "media/infracost.svg",
+          "contextualTitle": "Infracost Projects",
+          "when": "infracost:loggedIn"
+        }
+      ]
+    },
+    "viewsWelcome": [
+      {
+        "view": "infracostAuth",
+        "contents": "Welcome to Infracost for Visual Studio Code.🚀🚀🚀 \nLet's start by connecting VSCode with your Infracost Cloud account:\n[Connect VSCode to Infracost](command:infracost.login 'Connect with Infracost')"
+      }
+    ],
+    "commands": [
+      {
+        "command": "infracost.resourceBreakdown",
+        "title": "Show all the cost components for a given resource."
+      },
+      {
+        "command": "infracost.login",
+        "title": "Login to an Infracost Cloud account."
+      },
+      {
+        "command": "infracost.refresh",
+        "title": "Refresh",
+        "icon": {
+          "light": "resources/light/refresh.svg",
+          "dark": "resources/dark/refresh.svg"
+        }
+      }
+    ],
+    "menus": {
+      "view/title": [
+        {
+          "command": "infracost.refresh",
+          "when": "view == infracostProjects",
+          "group": "navigation"
+        }
+      ],
+      "view/item/context": []
+    }
+  },
+  "scripts": {
+    "vscode:package": "vsce package --yarn",
+    "vscode:prepublish": "npm run download:artifacts && yarn package",
+    "download:artifacts": "./scripts/download.sh",
+    "compile": "webpack",
+    "watch": "webpack --watch",
+    "package": "webpack --mode production --devtool hidden-source-map",
+    "lint": "eslint src --ext ts,js --ignore-path .eslintignore --ignore-path .gitignore . --max-warnings=0",
+    "lint:fix": "eslint src --fix --ext ts,js --ignore-path .eslintignore --ignore-path .gitignore . ",
+    "format": "prettier --write 'src/**/*.{js,ts}'",
+    "format:check": "prettier --check 'src/**/*.{js,ts}'"
+  },
+  "devDependencies": {
+    "@types/glob": "^7.2.0",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "14.x",
+    "@types/vscode": "^1.67.0",
+    "@typescript-eslint/eslint-plugin": "^5.30.5",
+    "@typescript-eslint/parser": "^5.30.5",
+    "@vscode/test-electron": "^2.1.3",
+    "esbuild": "^0.14.49",
+    "eslint": "^8.14.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "file-loader": "^6.2.0",
+    "glob": "^8.0.1",
+    "mocha": "^9.2.2",
+    "prettier": "^2.7.1",
+    "ts-loader": "^9.2.8",
+    "typescript": "^4.6.4",
+    "vsce": "^2.9.2",
+    "webpack": "^5.70.0",
+    "webpack-cli": "^4.9.2"
+  },
+  "dependencies": {
+    "@types/semver": "^7.3.12",
+    "handlebars": "^4.7.7",
+    "semver": "^7.3.8"
+  }
 }
