+++ src/cli.ts
@@ -1,0 +1,105 @@
+import { spawnSync, SpawnSyncReturns } from 'child_process';
+
+export namespace infracostJSON {
+  export interface Metadata {
+    path: string;
+    type: string;
+    vcsRepoUrl: string;
+    vcsSubPath: string;
+  }
+
+  export interface PastBreakdown {
+    resources: Resource[];
+    totalHourlyCost: string;
+    totalMonthlyCost: string;
+  }
+
+  export interface ResourceMetadata {
+    filename: string;
+    calls: Call[];
+  }
+
+  export interface Call {
+    blockName: string;
+    filename: string;
+  }
+
+  export interface CostComponent {
+    name: string;
+    unit: string;
+    hourlyQuantity: number;
+    monthlyQuantity: number;
+    price: string;
+    hourlyCost: number;
+    monthlyCost: number;
+  }
+
+  export interface Resource {
+    name: string;
+    metadata: ResourceMetadata;
+    hourlyCost: string;
+    monthlyCost: number;
+    costComponents: CostComponent[];
+    subresources: Resource[];
+  }
+
+  export interface Breakdown {
+    resources: Resource[];
+    totalHourlyCost: string;
+    totalMonthlyCost: string;
+  }
+
+  export interface Diff {
+    resources: Resource[];
+    totalHourlyCost: string;
+    totalMonthlyCost: string;
+  }
+
+  export interface Summary {
+    totalDetectedResources: number;
+    totalSupportedResources: number;
+    totalUnsupportedResources: number;
+    totalUsageBasedResources: number;
+    totalNoPriceResources: number;
+    unsupportedResourceCounts: Record<string, number>;
+    noPriceResourceCounts: Record<string, number>;
+  }
+
+  export interface Project {
+    name: string;
+    metadata: Metadata;
+    pastBreakdown: PastBreakdown;
+    breakdown: Breakdown;
+    diff: Diff;
+    summary: Summary;
+  }
+
+  export interface RootObject {
+    version: string;
+    currency: string;
+    projects: Project[];
+    totalHourlyCost: string;
+    totalMonthlyCost: string;
+    pastTotalHourlyCost: string;
+    pastTotalMonthlyCost: string;
+    diffTotalHourlyCost: string;
+    diffTotalMonthlyCost: string;
+    timeGenerated: Date;
+    summary: Summary;
+  }
+}
+
+export default class CLI {
+  constructor(private binaryPath: string) {}
+
+  async exec(...args: string[]): Promise<SpawnSyncReturns<Buffer>> {
+    return spawnSync(this.binaryPath, args, {
+      env: {
+        ...process.env,
+        INFRACOST_CLI_PLATFORM: 'vscode',
+        INFRACOST_NO_COLOR: 'true',
+        INFRACOST_SKIP_UPDATE_CHECK: 'true',
+      },
+    });
+  }
+}
